requirements.txt
# Minimal requirements for MNIST Digit Classifier
streamlit==1.26.0
numpy==1.23.5
Pillow==10.0.0
tensorflow==2.13.0
streamlit-drawable-canvas==0.9.2
matplotlib==3.7.2
